fix: research-intel schema alignment + DB write + stats casting
- source_papers: list[dict] → list[str] in Python schema (matches SQL TEXT[] and TS string[])
- Added source_claims field to Python schema (was in TS/SQL but missing in Python)
- Added --store flag to run_all.py — writes signals to research_signals table
- Supports psycopg2 or psql subprocess fallback for DB writes
- Stats endpoint: COUNT(*)::int to return numbers not strings
- gap_detector: _paper_ref returns ID string instead of {id, title} dict

+++ orchestrator/src/routes/research-intel.ts
@@ -73,10 +73,10 @@ export function researchIntelRoutes(pool: pg.Pool): Router {
     try {
       const { rows } = await pool.query(`
         SELECT
-          (SELECT COUNT(*) FROM research_signals WHERE batch_date = CURRENT_DATE) AS signals_today,
-          (SELECT COUNT(*) FROM research_opportunities WHERE batch_date = CURRENT_DATE) AS opportunities_today,
-          (SELECT COUNT(DISTINCT batch_date) FROM research_signals) AS total_batches,
-          (SELECT COUNT(*) FROM research_signals) AS total_signals
+          (SELECT COUNT(*)::int FROM research_signals WHERE batch_date = CURRENT_DATE) AS signals_today,
+          (SELECT COUNT(*)::int FROM research_opportunities WHERE batch_date = CURRENT_DATE) AS opportunities_today,
+          (SELECT COUNT(DISTINCT batch_date)::int FROM research_signals) AS total_batches,
+          (SELECT COUNT(*)::int FROM research_signals) AS total_signals
       `);
       res.json(rows[0] || {});
     } catch (err) {
